Fixed curry problems with keyword args/annotations on python 3
