+++ toolz/tests/test_functoolz.py
@@ -68,7 +68,6 @@ def test_memoize_kwargs():
 
 
 def test_memoize_curried():
-    @curry
     def f(x, y=0):
         return x + y
